Move downsampling to camera driver

See merge request robotics/trifinger_object_tracking!77

+++ src/pose_detector.cpp
@@ -185,8 +185,9 @@ float PoseDetector::cost_function(
              color_idx < dominant_colors[camera_idx].size();
              color_idx++)
         {
-            unsigned int num_pixels = masks_pixels[camera_idx][color_idx].size();
-            if(num_pixels == 0)
+            unsigned int num_pixels =
+                masks_pixels[camera_idx][color_idx].size();
+            if (num_pixels == 0)
             {
                 continue;
             }
@@ -283,7 +284,7 @@ void PoseDetector::optimize_using_optim(
     const std::array<std::vector<FaceColor>, N_CAMERAS> &dominant_colors,
     const std::array<std::vector<cv::Mat>, N_CAMERAS> &masks)
 {
-    //ScopedTimer timer("PoseDetector/optim");
+    // ScopedTimer timer("PoseDetector/optim");
 
     std::array<std::vector<std::vector<cv::Point>>, N_CAMERAS> masks_pixels;
     for (int camera_idx = 0; camera_idx < N_CAMERAS; camera_idx++)
@@ -371,14 +372,15 @@ void PoseDetector::optimize_using_optim(
                   &num_misclassified_pixels_);
 
     // TODO: Consider number of pixels in the segmentation
-    confidence_ = 1 - num_misclassified_pixels_ / num_samples;
+    confidence_ = 1 - static_cast<float>(num_misclassified_pixels_) /
+                          static_cast<float>(num_samples);
 }
 
 Pose PoseDetector::find_pose(
     const std::array<std::vector<FaceColor>, N_CAMERAS> &dominant_colors,
     const std::array<std::vector<cv::Mat>, N_CAMERAS> &masks)
 {
-    //ScopedTimer timer("PoseDetector/find_pose");
+    // ScopedTimer timer("PoseDetector/find_pose");
 
     // calculates mean_position and mean_orientation
     optimize_using_optim(dominant_colors, masks);
